refactor: move ranking service into service folder
