+++ integration/service/ranking/flags.go
@@ -1,4 +1,4 @@
-package rankingservice
+package ranking
 
 import "flag"
