+++ integration/service/ranking/route_by_osrm.go
@@ -1,4 +1,4 @@
-package rankingservice
+package ranking
 
 import (
 	"encoding/json"
