+++ integration/service/ranking/update_by_traffic.go
@@ -1,4 +1,4 @@
-package rankingservice
+package ranking
 
 import (
 	"math"
